+++ website/src/Pages/About/About.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import Page1 from './Page1'
+
+
+const About = () => {
+  return <>
+  <Page1 />
+  </>
+}
+
+export default About
