+++ website/src/Pages/About/Page1.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Page1 = () => {
+  return (
+    <div>Page1</div>
+  )
+}
+
+export default Page1
