+++ website/src/Pages/Careers/Careers.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Careers = () => {
+  return (
+    <div>Careers</div>
+  )
+}
+
+export default Careers
