+++ website/src/Pages/Contacts/Contact.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Contact = () => {
+  return (
+    <div>Contact</div>
+  )
+}
+
+export default Contact
