+++ website/src/Pages/Home/Home.js
@@ -1,0 +1,26 @@
+import React from 'react'
+import Page1 from './Page1'
+import Page2 from './Page2'
+import Page3 from './Page3'
+import Page4 from './Page4'
+import Page5 from './Page5'
+import India from './India'
+import Page7 from './Page7'
+import Page9 from './Page9'
+import Footer from '../../Components/Footer'
+
+const Home = () => {
+  return <>
+  <Page1 />
+  <Page2 />
+  <Page3 />
+  <Page4 />
+  <Page5 />
+  <India />
+  <Page7 />
+  <Page9 />
+  <Footer />
+  </>
+}
+
+export default Home
