+++ website/src/Pages/Home/Page9.js
@@ -3,35 +3,35 @@ import { Box, Typography, Avatar } from "@mui/material";
 import { useTheme, useMediaQuery } from "@mui/material";
 
 import { FaQuoteLeft } from "react-icons/fa";
-import Img1 from "../assets/client_logo/1.png";
-import Img2 from "../assets/client_logo/2.png";
-import Img3 from "../assets/client_logo/3.png";
-import Img4 from "../assets/client_logo/4.png";
-import Img5 from "../assets/client_logo/5.png";
-import Img6 from "../assets/client_logo/6.png";
-import Img7 from "../assets/client_logo/7.png";
-import Img8 from "../assets/client_logo/8.png";
-import Img9 from "../assets/client_logo/9.png";
-import Img10 from "../assets/client_logo/10.png";
-import Img11 from "../assets/client_logo/11.png";
-import Img12 from "../assets/client_logo/12.png";
-import Img13 from "../assets/client_logo/13.png";
-import Img14 from "../assets/client_logo/14.png";
-import Img15 from "../assets/client_logo/15.png";
-import Img16 from "../assets/client_logo/16.png";
-import Img17 from "../assets/client_logo/17.png";
-import Img18 from "../assets/client_logo/18.png";
-import Img19 from "../assets/client_logo/19.png";
-import Img20 from "../assets/client_logo/20.png";
-import Img21 from "../assets/client_logo/21.png";
-import Img22 from "../assets/client_logo/22.png";
-import Img23 from "../assets/client_logo/23.png";
-import Img24 from "../assets/client_logo/24.png";
-import Img25 from "../assets/client_logo/25.png";
-import Img26 from "../assets/client_logo/26.png";
-import Img27 from "../assets/client_logo/27.png";
-import Img28 from "../assets/client_logo/28.png";
-import Img29 from "../assets/client_logo/29.png";
+import Img1 from "../../assets/client_logo/1.png";
+import Img2 from "../../assets/client_logo/2.png";
+import Img3 from "../../assets/client_logo/3.png";
+import Img4 from "../../assets/client_logo/4.png";
+import Img5 from "../../assets/client_logo/5.png";
+import Img6 from "../../assets/client_logo/6.png";
+import Img7 from "../../assets/client_logo/7.png";
+import Img8 from "../../assets/client_logo/8.png";
+import Img9 from "../../assets/client_logo/9.png";
+import Img10 from "../../assets/client_logo/10.png";
+import Img11 from "../../assets/client_logo/11.png";
+import Img12 from "../../assets/client_logo/12.png";
+import Img13 from "../../assets/client_logo/13.png";
+import Img14 from "../../assets/client_logo/14.png";
+import Img15 from "../../assets/client_logo/15.png";
+import Img16 from "../../assets/client_logo/16.png";
+import Img17 from "../../assets/client_logo/17.png";
+import Img18 from "../../assets/client_logo/18.png";
+import Img19 from "../../assets/client_logo/19.png";
+import Img20 from "../../assets/client_logo/20.png";
+import Img21 from "../../assets/client_logo/21.png";
+import Img22 from "../../assets/client_logo/22.png";
+import Img23 from "../../assets/client_logo/23.png";
+import Img24 from "../../assets/client_logo/24.png";
+import Img25 from "../../assets/client_logo/25.png";
+import Img26 from "../../assets/client_logo/26.png";
+import Img27 from "../../assets/client_logo/27.png";
+import Img28 from "../../assets/client_logo/28.png";
+import Img29 from "../../assets/client_logo/29.png";
 
 const clients = [
   Img1, Img2, Img3, Img4, Img5, Img6, Img7, Img8, Img9, Img10,
