+++ website/src/Pages/Kusum Yojana/Kusum.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Kusum = () => {
+  return (
+    <div>Kusum</div>
+  )
+}
+
+export default Kusum
