+++ website/src/Pages/Services/Services.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Services = () => {
+  return (
+    <div>Services</div>
+  )
+}
+
+export default Services
